fix: switch from managing the open and close state of the menu using state
an attemtp to use state to manage the opening and closing of the
navigation drawer led to a subtle bug.
when the website is opened on desktop and resized to a mobile screen,
the navigation menu doesn't automatically go away.
it required the use to manually click the hamburger icon to hide it.
a more lightweight and concise solution was achieved using pure css